rev1.0.3 * svg renderer added * json renderer moved to render/json.js * xml renerer uses HTML one temporary * README.md updated to reflect ^^
