+++ demo/index.html
@@ -52,8 +52,12 @@
 <html-demo-element legend="array dwarfs.json as table">
     <slotted-element src="dwarfs.json"></slotted-element>
 </html-demo-element>
+
+<html-demo-element legend="src is image confused.svg">
+    <slotted-element src="confused.svg"></slotted-element>
+</html-demo-element>
 </main>
-<script type="module" src="https://unpkg.com/html-demo-element@1.0.5/html-demo-element.js"></script>
+<script type="module" src="https://unpkg.com/html-demo-element@1.0.7/html-demo-element.js"></script>
 <script type="module" src="../slotted-element.js"></script>
 
 </body>
