+++ demo/render-from-xml.html
@@ -38,9 +38,5 @@
         <fetch-element src="dwarfs.json"></fetch-element>
     </html-demo-element>
 </main>
-<main>
-    <html-demo-element legend="dwarfs.json" src="dwarfs.json"></html-demo-element>
-    <html-demo-element legend="doc.json" src="doc.json"></html-demo-element>
-</main>
 </body>
 </html>
