+++ demo/template-demo.html
@@ -5,7 +5,7 @@
     <style>
         @import "demo.css";
     </style>
-    <script type="module" src="https://unpkg.com/html-demo-element@1.0.5/html-demo-element.js"></script>
+    <script type="module" src="https://unpkg.com/html-demo-element@1.0.7/html-demo-element.js"></script>
     <script type="module" src="../slotted-element.js"></script>
 
 </head>
