+++ fetch-element.js
@@ -20,6 +20,9 @@ FetchElement extends HTMLElement
 
     static get mime2mod(){ return   {   'application/json':'./render/json.js'
                                     ,   'text/html': FetchElement.prototype.renderHtml
+                                    ,   'text/xml': FetchElement.prototype.renderHtml
+                                    ,   'application/xml': FetchElement.prototype.renderHtml
+                                    ,   'image/svg+xml': FetchElement.prototype.renderHtml
                                     }}
 
     get headers(){ return {} }
@@ -140,39 +143,6 @@ FetchElement extends HTMLElement
 
     getKeys( obj ){ return Object.keys( obj ); }
 
-    json2table( data, path )
-    {
-        if( Array.isArray( data ) )
-        {
-            if( !data.length )
-                return '';
-            const keys = this.getKeys( data[ 0 ] );
-
-            return `
-<table>
-    <tr>${ keys.map( k => `<th>${ k }</th>` ).join( '\n' ) }</tr>
-    ${ data.map( ( r, i ) => `
-    <tr>${ keys.map( k => `
-        <td key="${ toKebbabCase( k ) }">
-            ${ this.json2table( r[ k ], [ ...path, i, k ] ) }
-        </td>` ).join( '' )
-            }
-    </tr>` ).join( '\n' ) }
-</table>
-`
-        }
-        if( typeof data !== 'object' || data === null )
-            return data;
-        const keys = this.getKeys( data );
-        return `
-<table>
-    ${ keys.map( k => `
-<tr><th>${ k }</th>
-    <td key="${ toKebbabCase( k ) }">${ this.json2table( data[ k ], [ ...path, k ] ) }</td>
-</tr>` ).join( '' ) }
-</table>
-`
-    }
 
 }
